further organize functions into 'admin' and 'code' modules
- reduces dependencies in user code
